move(external): 알림 관련 기능 internal -> external 로 변경
