+++ main/src/main/java/org/sopt/makers/crew/main/external/notification/PushNotificationEnums.java
@@ -1,4 +1,4 @@
-package org.sopt.makers.crew.main.internal.notification;
+package org.sopt.makers.crew.main.external.notification;
 
 import lombok.AccessLevel;
 import lombok.Getter;
