+++ main/src/main/java/org/sopt/makers/crew/main/external/notification/dto/PushNotificationRequestDto.java
@@ -1,4 +1,4 @@
-package org.sopt.makers.crew.main.internal.notification.dto;
+package org.sopt.makers.crew.main.external.notification.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
