+++ main/src/main/java/org/sopt/makers/crew/main/external/notification/dto/PushNotificationResponseDto.java
@@ -1,4 +1,4 @@
-package org.sopt.makers.crew.main.internal.notification.dto;
+package org.sopt.makers.crew.main.external.notification.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
